Hug update Dashboard

+++ src/main/java/com/dhakcare/dao/DoctorDAO.java
@@ -4,6 +4,6 @@ import com.dhakcare.entity.Doctor;
 import com.dhakcare.utils.GenericDAO;
 
 public interface DoctorDAO extends GenericDAO<Doctor>{
-	public Integer countTotalDoctor();
+	public Long countTotalDoctor();
   
 }
