+++ src/main/java/com/dhakcare/service/PatientService.java
@@ -7,6 +7,6 @@ import com.dhakcare.entity.Patient;
 public interface PatientService {
 	public boolean createPatient(Patient patient);
 	public List<Patient> findPatientbyUserId(Long id);
-	public Integer getTotalPatient();
+	public Long getTotalPatient();
 	public boolean deleteByUserId(String id);
 }
